feat: refactor invites management and introduce InviteCard component
